Update blackjack.asm
Added "money betting" functions, using the float registers to manage this kind of values. Also found a bug (inherited from last update) where the player's sum of points gets to zero and therefore makes the comparison between the user's and the dealer's points wrong, but I haven't been able to fix it yet.